+++ frontend/app/(droplets)/d/[slug]/recap/confetti.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-/*
 import { useEffect } from "react";
 import confetti from "canvas-confetti";
 
@@ -63,4 +62,3 @@ export function Confetti() {
   return null;
 }
 
-*/
