+++ frontend/lib/requests/authorized-user.ts
@@ -1,3 +1,4 @@
+"use server"
 import { fetchAPI, flattenAttributes } from "@/lib/utils";
 import { AuthorizedUser } from "@/types";
 import { StrapiRequestParams } from "@/types/strapi";
